Change standard C++ code to Unreal programming model

+++ Private/Chunk/Main.cpp
@@ -2,13 +2,13 @@
 
 #include "Chunk/Main.h"
 
-FVoxChunkMain FVoxChunkMain::Read(const void*& data, size_t& size)
+FVoxChunkMain FVoxChunkMain::Read(const void*& data, int64& size)
 {
 	auto cmain = FVoxChunkMain();
-	auto id = ReadData<uint32_t>(data, size);
+	auto id = ReadData<uint32>(data, size);
 	if (FVoxChunkMain::Tag != id) return cmain;
 
-	cmain.Content = ReadData<int32_t>(data, size);
-	cmain.Children = ReadData<int32_t>(data, size);
+	cmain.Content = ReadData<int32>(data, size);
+	cmain.Children = ReadData<int32>(data, size);
 	return cmain;
 }
