+++ Private/Chunk/Pack.cpp
@@ -2,14 +2,14 @@
 
 #include "Chunk/Pack.h"
 
-FVoxChunkPack FVoxChunkPack::Read(const void*& data, size_t& size)
+FVoxChunkPack FVoxChunkPack::Read(const void*& data, int64& size)
 {
 	auto cpack = FVoxChunkPack();
-	auto id = ReadData<uint32_t>(data, size);
+	auto id = ReadData<uint32>(data, size);
 	if (FVoxChunkPack::Tag != id) return cpack;
 
-	cpack.Content = ReadData<int32_t>(data, size);
-	cpack.Children = ReadData<int32_t>(data, size);
-	cpack.Models = ReadData<int32_t>(data, size);
+	cpack.Content = ReadData<int32>(data, size);
+	cpack.Children = ReadData<int32>(data, size);
+	cpack.Models = ReadData<int32>(data, size);
 	return cpack;
 }
