+++ Private/Chunk/Size.cpp
@@ -2,16 +2,16 @@
 
 #include "Chunk/Size.h"
 
-FVoxChunkSize FVoxChunkSize::Read(const void*& data, size_t& size)
+FVoxChunkSize FVoxChunkSize::Read(const void*& data, int64& size)
 {
 	auto csize = FVoxChunkSize();
-	auto id = ReadData<uint32_t>(data, size);
+	auto id = ReadData<uint32>(data, size);
 	if (FVoxChunkSize::Tag != id) return csize;
 
-	csize.Content = ReadData<int32_t>(data, size);
-	csize.Children = ReadData<int32_t>(data, size);
-	csize.X = ReadData<int32_t>(data, size);
-	csize.Y = ReadData<int32_t>(data, size);
-	csize.Z = ReadData<int32_t>(data, size);
+	csize.Content = ReadData<int32>(data, size);
+	csize.Children = ReadData<int32>(data, size);
+	csize.X = ReadData<int32>(data, size);
+	csize.Y = ReadData<int32>(data, size);
+	csize.Z = ReadData<int32>(data, size);
 	return csize;
 }
