+++ Public/Chunk/Size.h
@@ -5,16 +5,14 @@
 #include "CoreMinimal.h"
 #include "Chunk.h"
 
-#include <iostream>
-
 /**
  * Size chunk
  */
 struct FVoxChunkSize : public FVoxChunk
 {
-	static constexpr uint32_t Tag = GenerateId('S', 'I', 'Z', 'E');
+	static constexpr uint32 Tag = GenerateId('S', 'I', 'Z', 'E');
 
-	int32_t X, Y, Z;
+	int32 X, Y, Z;
 
-	static FVoxChunkSize Read(const void*& data, size_t& size);
+	static FVoxChunkSize Read(const void*& data, int64& size);
 };
